Fixed all comments W1-D3-View and Group View
