+++ app/src/at-phutran/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ViewActivity"
+        <activity android:name=".ViewAndGroupView.ViewActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"></activity>
-        <activity android:name=".ListActivity">
+        <activity android:name=".ViewAndGroupView.ListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
