+++ app/src/at-phutran/java/asiantech/internship/summer/ViewAndGroupView/ListActivity.java
@@ -1,4 +1,4 @@
-package asiantech.internship.summer;
+package asiantech.internship.summer.ViewAndGroupView;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.Window;
 import android.view.WindowManager;
 import android.widget.Button;
+
+import asiantech.internship.summer.R;
 
 public class ListActivity extends AppCompatActivity {
