+++ app/src/at-phutran/res/layout-v16/activity_view.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewActivity">
+    tools:context=".ViewAndGroupView.ViewActivity">
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="110dp"
