+++ app/src/at-phutran/res/layout/activity_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListActivity"
+    tools:context=".ViewAndGroupView.ListActivity"
     android:orientation="vertical"
     android:gravity="center">
     <Button
