+++ app/src/at-phutran/res/layout/activity_view.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ViewActivity">
+    tools:context=".ViewAndGroupView.ViewActivity">
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="110dp"
